chore: git hygiene — commit missing infra/docs, tighten .gitignore (#46)
* chore: git hygiene — commit missing infra/docs, tighten .gitignore

Four untracked items that have been sitting in working tree for days:

- infra/ansible/roles/{app,caddy,common,ollama,postgres}/handlers/main.yml
  — five Ansible handler files referenced by their roles' tasks.
  Without them, a fresh checkout would fail to deploy.

- docs/telegram-credentials-setup.md — onboarding doc walking through
  api_id/api_hash creation at my.telegram.org and bot setup with
  @BotFather. All credential values shown are obvious placeholders
  (1234567, abcdef…, AAF_abc123xyz…).

- scripts/dev/list-chats.py — moved from repo root (was test_telegram.py).
  Dev utility for sanity-checking Telethon credentials by listing top 20
  chats and dumping recent messages from one. Renamed for clarity and
  relocated to fit the existing scripts/ convention.

- .gitignore: replace `.claude/settings.local.json` with `.claude/`
  to ignore the whole local agent state directory (settings, worktrees,
  scratch).

Also cleaned locally (not part of this commit):
- .git/REBASE_HEAD orphan pointer (no rebase was in progress).


* fix(types): clear strict-mypy errors blocking main CI

Main's CI has been red since 1ea1fde. Six errors split across three files,
all mypy strict-mode complaints — no behavior changes. Plus two trivial
errors in scripts/dev/list-chats.py from this PR's earlier commit.

- worker-chat-tagger/centroids.py:_cosine — wrap division in float()
  (sum() of generator expression returns Any in strict).
- worker-chat-tagger/classifier.py:_parse_llm_json — annotate dict[str, Any]
  return type and intermediate variables to stop json.loads Any leakage.
- worker-chat-tagger/classifier.py:_ollama_chat — bind data["message"]
  ["content"] to a str-typed local before returning.
- mcp-server/main.py:_handle_tool_call — rename `result` to `brief` in the
  get_recent_brief branch so it doesn't collide with the CommitmentResult
  type inferred from earlier reassignments.
- scripts/dev/list-chats.py:main — add `-> None` annotation.


---------

+++ services/worker-chat-tagger/tbc_worker_chat_tagger/centroids.py
@@ -59,7 +59,7 @@ def _cosine(a: list[float], b: list[float]) -> float:
     nb = sum(y * y for y in b) ** 0.5
     if na == 0 or nb == 0:
         return 0.0
-    return dot / (na * nb)
+    return float(dot / (na * nb))
 
 
 def _fetch_chat_embeddings(
